made ThemeToggleButton independent component
