+++ src/components/layout.js
@@ -5,6 +5,7 @@ import '../styles/layout.css'
 import { pageGrid, breakpoint, colGap, rowGap } from '../styles/sizes'
 import Header from './Header'
 import GlobalStyle from '../styles/GlobalStyle'
+import ThemeToggleButton from './ThemeToggleButton'
 
 const Grid = styled.div`
 	--header-row: calc(${rowGap} * 2.75);
@@ -90,7 +91,7 @@ const Layout = props => {
 	}
 	return (
 		<ThemeProvider theme={theme.current}>
-			<Header themetogglebutton={<theme.ToggleButton />} />
+			<Header themetogglebutton={<ThemeToggleButton settheme={theme.setTheme} />} />
 			<Grid>{props.children}</Grid>
 			<GlobalStyle />
 		</ThemeProvider>
