+++ src/hooks/useTheme.js
@@ -1,54 +1,9 @@
-import React, { useState, useEffect, useMemo, useCallback } from 'react'
+import { useState, useEffect, useCallback } from 'react'
 import themes from '../styles/themes'
-import styled from 'styled-components'
-import { useSpring, animated, config } from 'react-spring'
 
-const NonNavigatingButton = styled.div`
-	cursor: pointer;
-`
+
 const useTheme = () => {
 	const [current, setCurrent] = useState(null)
-	const [hovering, setHovering] = useState(false)
-
-	const themeIconProps = useMemo(
-		() => ({
-			color: hovering
-				? current?.purpleTextStrong || 'blue'
-				: current?.purpleText || 'purple',
-			rotation: `rotate(${hovering ? 60 : 25})`,
-			config: config.default
-		}),
-		[hovering, current]
-	)
-
-	const lightIconProps = {
-		...themeIconProps,
-		name: 'light',
-		cx0: 12.68,
-		cy0: 12.68,
-		r0: 11.18,
-		strokeWidth: 3,
-		strokeDasharray: '2.37 4.75',
-		cx1: 12.68,
-		cy1: 12.68,
-		r1: 7.68
-	}
-	const darkIconProps = {
-		...themeIconProps,
-		name: 'dark',
-		cx0: 13.0405,
-		cy0: 12.6802,
-		r0: 9.417,
-		strokeWidth: 2.527,
-		strokeDasharray: '7 0',
-		cx1: 8.959,
-		cy1: 12.6802,
-		r1: 6.68
-	}
-
-	const [themeButtonProps, api] = useSpring(() =>
-		current?.name === 'dark' ? darkIconProps : lightIconProps
-	)
 
 	const systemPrefersDarkTheme =
 		typeof window !== 'undefined'
@@ -93,14 +48,6 @@ const useTheme = () => {
 	)
 
 	useEffect(() => {
-		api.start({
-			...themeIconProps,
-			config: {
-				mass: 1,
-				tension: 400,
-				friction: 15
-			}
-		})
 
 		if (!current) {
 			if (storageAvailable()) {
@@ -145,8 +92,6 @@ const useTheme = () => {
 			} else setCurrent(themes['light'])
 		}
 	}, [
-		api,
-		themeIconProps,
 		systemPrefersDarkTheme,
 		current,
 		setThemeAgent,
@@ -154,79 +99,15 @@ const useTheme = () => {
 		setStorage
 	])
 
-	const toggleTheme = () => {
-		const newTheme = current?.name === 'dark' ? themes.light : themes.dark
-		const newIcon =
-			themeButtonProps.name.get() === 'light' ? darkIconProps : lightIconProps
+	const setTheme = themeName => {
+		const newTheme = themes[themeName]
 		setStorage('theme', [newTheme.name, setThemeAgent(newTheme.name)])
-		api.start({ to: newIcon })
 		setCurrent(newTheme)
 	}
 
-	const themeIconEvent = e => {
-		switch (e.type) {
-			case 'mouseenter':
-				setHovering(true)
-				break
-			case 'mouseleave':
-				setHovering(false)
-				break
-			case 'keydown':
-				if (e.key !== 'Enter') {
-					break
-				}
-				toggleTheme()
-				break
-			case 'click':
-				toggleTheme()
-				break
-			default:
-				console.log('unhandled mouse event:', e.type)
-		}
-	}
-
-	const ToggleButton = (props) => (
-		<NonNavigatingButton
-			{...props}
-			tabIndex='0'
-			aria-label='toggle dark mode'
-			id='theme-toggle-button'
-			role='button'
-			onKeyDown={e => themeIconEvent(e)}
-			onClick={e => themeIconEvent(e)}
-			onMouseLeave={e => themeIconEvent(e)}
-			onMouseEnter={e => themeIconEvent(e)}
-		>
-			<animated.svg
-				role='document'
-				xmlns='http://www.w3.org/2000/svg'
-				width={26}
-				height={26}
-				fill='none'
-				transform={themeButtonProps.rotation}
-			>
-				<animated.circle
-					cx={themeButtonProps.cx0}
-					cy={themeButtonProps.cy0}
-					r={themeButtonProps.r0}
-					stroke={themeButtonProps.color}
-					strokeWidth={themeButtonProps.strokeWidth}
-					strokeDasharray={themeButtonProps.strokeDasharray}
-				/>
-				<animated.circle
-					cx={themeButtonProps.cx1}
-					cy={themeButtonProps.cy1}
-					r={themeButtonProps.r1}
-					fill={themeButtonProps.color}
-				/>
-			</animated.svg>
-		</NonNavigatingButton>
-	)
-
 	return {
 		current,
-		ToggleButton,
-		toggleTheme
+		setTheme,
 	}
 }
